Moved to Unity 2021 LTS + Bunch-o-content

+++ Assets/Source/Enemies/Waves/Generators/GeneratorEnemyData.cs
@@ -11,7 +11,7 @@ namespace Lomztein.BFA2
         [ModelProperty]
         public string EnemyIdentifier;
         [ModelProperty]
-        public float DifficultyValue;
+        public double DifficultyValue;
         [ModelProperty]
         public int EarliestWave;
         [ModelProperty]
